requirements.txt
supabase>=2.0.0
tiktoken>=0.6.0
httpx>=0.27.0
python-dotenv>=1.0.0
networkx>=3.2
